requirements.txt
asgiref                      3.4.1
atomicwrites                 1.4.0
attrs                        21.4.0
bootstrap4                   0.1.0
click                        8.0.3
cmake                        3.22.2
colorama                     0.4.4
dataclasses                  0.8
DateTime                     4.4
Deprecated                   1.2.13
Django                       3.2.12
django-binary-database-files 1.0.15
django-bootstrap-form        3.4
django-crispy-forms          1.13.0
django-picklefield           3.0.1
djongo                       1.3.6
dlib                         19.8.1
et-xmlfile                   1.1.0
face-recognition             1.3.0
face-recognition-models      0.3.0
importlib-metadata           4.8.3
iniconfig                    1.1.1
joblib                       1.1.0
jsons                        1.6.1
mongo-wrapper                1.0.0
numpy                        1.19.5
opencv-contrib-python        4.5.5.62
opencv-python                4.5.4.60
openpyxl                     3.0.9
packaging                    21.3
pandas                       1.1.5
Pillow                       8.4.0
pip                          21.3.1
pluggy                       1.0.0
py                           1.11.0
pymongo                      3.12.1
pyparsing                    3.0.7
pytest                       7.0.0
python-dateutil              2.8.2
pytz                         2021.3
redis                        4.1.3
scikit-learn                 0.24.2
scipy                        1.5.4
setuptools                   40.6.2
six                          1.16.0
sqlparse                     0.2.4
threadpoolctl                3.1.0
tomli                        1.2.3
typing_extensions            4.0.1
typish                       1.9.3
wrapt                        1.13.3
zipp                         3.6.0
zope.interface               5.4.0